refactor: Move IAM files back to old location to avoid breaking change
